refactored code to use service
