+++ src/app/sharedComponents/recipes-display/recipe-card.ts
@@ -1,0 +1,5 @@
+export class RecipeCard {
+  title!: string;
+  url!: string;
+  image!: string;
+}
