+++ src/app/sharedComponents/recipes-display/recipe.service.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@angular/core';
+
+import { RecipeCard } from './recipe-card';
+import { RECIPECARDS } from './mock-recipes'
+@Injectable()
+
+export class RecipeService {
+  getRecipes(): Promise<RecipeCard[]> {
+    return Promise.resolve(RECIPECARDS);
+  }
+}
